// rsclone/src/modules/core/types/categoriesTypes.ts
type CategoriesType = {
  episodeList: string;
  episodeDetails: string;
  intro: string;
  seasonOne: string;
  seasonTwo: string;
  backToMain: string;
  startEpisode: string;
  hintTitle: string;
  minimumRating: string;
  duration: string;
};

export type CategoriesByLanguageType = {
  ru: CategoriesType;
  en: CategoriesType;
};